refactor(auth): server-driven quotaAutoSwitched + spendingStatus module
Polish the relay-quota wiring before merge:

- Pull SpendingStatusSchema into its own src/shared/schemas/spendingStatus.ts
  so both profile messages and auth/tier can import without cycles. tier/types
  re-exports for back-compat.
- Add wasQuotaAutoSwitched() on ServerSideKeyService and ship the boolean
  to the webview as quotaAutoSwitched in UPDATE_PROFILE instead of
  client-side diffing the remaining-token transition. Simpler dispatcher,
  no stale-flag risk across reloads.
- ClearServerSideKeyCachesOptions { resetQuotaFlip } so sign-in/out and
  manual toggle changes can choose whether to re-arm the auto-flip guard.
- Sign-in/out hooks across SupabaseAuthProvider, packages/cli, and
  packages/desktop now use resetQuotaFlip:true so a new session
  re-evaluates quota fresh.
