+++ packages/extension/src/settingsView/frontend/SettingsApp.ts
@@ -30,7 +30,7 @@ import {
   SETTINGS_TAB,
   SETTINGS_TAB_ORDER,
 } from '@shared/schemas';
-import type { SpendingStatus } from '@shared/schemas/profileViewMessages';
+import type { SpendingStatus } from '@shared/schemas/spendingStatus';
 import {
   registerTeXRAWebAwesomeIcons,
   TEXRA_ICON_LIBRARY,
@@ -255,12 +255,6 @@ export class SettingsApp extends SettingsAppBase {
   private readonly apiAccessMode = signal<'included' | 'personal'>('personal');
   private readonly spendingStatus = signal<SpendingStatus | null>(null);
   private readonly quotaAutoSwitched = signal(false);
-  // Mutable ref shared with the dispatcher so it can detect the
-  // remaining > 0 → remaining <= 0 transition across two UPDATE_PROFILE
-  // messages without exposing a signal for this purely internal state.
-  private readonly prevSpendRemainingRef: { current: number | null } = {
-    current: null,
-  };
   private readonly allowedModels = signal<string[] | null>([]);
   private readonly providerKeyStatuses = signal<ProviderKeyStatus[]>([]);
   private readonly globalStreamingDefault = signal(true);
@@ -345,7 +339,6 @@ export class SettingsApp extends SettingsAppBase {
       apiAccessMode: this.apiAccessMode,
       spendingStatus: this.spendingStatus,
       quotaAutoSwitched: this.quotaAutoSwitched,
-      prevSpendRemainingRef: this.prevSpendRemainingRef,
       allowedModels: this.allowedModels,
       providerKeyStatuses: this.providerKeyStatuses,
       globalStreamingDefault: this.globalStreamingDefault,
